Move the errors package into the compiler package.
