+++ internal/vm/compiler/codegen/codegen.go
@@ -14,8 +14,8 @@ import (
 	"github.com/google/mtail/internal/metrics/datum"
 	"github.com/google/mtail/internal/vm/code"
 	"github.com/google/mtail/internal/vm/compiler/ast"
+	"github.com/google/mtail/internal/vm/compiler/errors"
 	"github.com/google/mtail/internal/vm/compiler/parser"
-	"github.com/google/mtail/internal/vm/errors"
 	"github.com/google/mtail/internal/vm/object"
 	"github.com/google/mtail/internal/vm/position"
 	"github.com/google/mtail/internal/vm/symbol"
