+++ internal/vm/compiler/errors/errors.go
