+++ internal/vm/compiler/parser/driver.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/golang/glog"
 	"github.com/google/mtail/internal/vm/compiler/ast"
-	"github.com/google/mtail/internal/vm/errors"
+	"github.com/google/mtail/internal/vm/compiler/errors"
 	"github.com/google/mtail/internal/vm/position"
 )
